OpenCL Port
git-svn-id: https://mediacenter-pc/svn/GPGPU/CUDA/RaytracingEngine/trunk@1138 c4de0f11-4f5c-764a-a4aa-d9e147179871

+++ FileMarshaller.cpp
@@ -268,7 +268,7 @@ void FileMarshaller::readMaterial( GPUKernel& kernel, const std::string& line, c
             case 21: material.textureMapping.y    = static_cast<int>(atoi( value.c_str() )); break;
             case 22: material.textureMapping.z    = static_cast<int>(atoi( value.c_str() )); break;
             case 23: material.textureMapping.w    = static_cast<int>(atoi( value.c_str() )); break;
-            case 24: material.textureOffset.x     = static_cast<float>(atof( value.c_str() )); break;
+            case 24: material.textureOffset.x     = static_cast<int>(atoi( value.c_str() )); break;
             case 25: material.textureOffset.x     = static_cast<int>(atoi( value.c_str() )); break;
             case 26: material.textureOffset.y     = static_cast<int>(atoi( value.c_str() )); break;
             case 27: material.textureOffset.w     = static_cast<int>(atoi( value.c_str() )); break;
